change to 32

+++ scripts/model.py
@@ -106,7 +106,7 @@ class Model:
                     load_safety_checker=False,
                     controlnet=controlnet,
                     #local_files_only=True,
-                    torch_dtype=torch.float16 if self.device.type == 'cuda' else torch.float32)
+                    torch_dtype=torch.float32 if self.device.type == 'cuda' else torch.float32)
             
             else:
                 pipe = StableDiffusionControlNetPipeline.from_single_file(
